refactoring index.js
