+++ index.html
@@ -41,7 +41,7 @@
     </div>
 
     <!-- initialize your application -->
-    <script type="module" src="./src/init/index.js"></script>
+    <script type="module" src="./src/init.js"></script>
 
     <!-- load the review buttons for your live deployment  -->
     <script type="module" src="./lib/review-buttons.js"></script>
