+++ src/handlers/showResults.js
@@ -1,0 +1,5 @@
+'use strict';
+
+export function showResults() {
+    window.location.replace("./review.html");
+}
