+++ src/logic/allAnswers.js
@@ -1,0 +1,10 @@
+'use strict';
+
+import { quizData } from "../data.js";
+
+// correct answers all in one array
+export const correctAnswers = [];
+for (let question of quizData.questions) {
+    correctAnswers.push(question.answers[question.correct]);
+}
+console.table(correctAnswers);
