+++ src/logic/reviewQuizSetter.js
@@ -1,0 +1,8 @@
+'use strict';
+
+import { showResults } from '../handlers/showResults.js';
+
+export const reviewQuizSetter = (id) => {
+    const endQuizBtn = document.getElementById(id);
+    endQuizBtn.addEventListener('click', showResults);
+}
